pyproject.toml
[project]
name = "mlflow-dock"
version = "0.1.0"
description = "MLflow webhook receiver that automatically builds and pushes Docker images when new model versions are created"
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Louis MOULIN", email = "moulin.louis@protonmail.com"}
]
keywords = ["mlflow", "docker", "webhook", "machine-learning", "mlops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "celery>=5.6.2",
    "docker>=7.1.0",
    "fastapi[standard]>=0.128.0",
    "mlflow>=3.8.1",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
mlflow-dock = "mlflow_dock.main:main"
create-webhook = "scripts.create_webhook:main"

[project.urls]
Homepage = "https://github.com/moulin-louis/mlflow-dock"
Repository = "https://github.com/moulin-louis/mlflow-dock"
Issues = "https://github.com/moulin-louis/mlflow-dock/issues"

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
